Allow the user to specify an alternate editor via $EDITOR.
Not everyone uses vim, and it's reasonable to check for $EDITOR
before forcing them to use vim. Ideally I think that we should
check for $EDITOR (and possibly a couple of others) and then poll
the user for a choice if that's not set, but this will do for now.